API

MuTuner.MuTunerLogger

MuTuner.MuTunerLogger

MuTuner.replay

MuTuner.update!

MuTuner.update_forgetful_mean

MuTuner.update_forgetful_mv

MuTuner.update_forgetful_var

MuTuner.MuTunerLogger — Type

MuTunerLogger{T<:AbstractFloat}

Type to assist with dynamically tuning the chemical potential $\mu$ to achieve a target particle density $\langle n \rangle$.

Fields

n₀::AbstractFloat

Target filling as intensive value $\langle n \rangle$.

N₀::AbstractFloat

Target filling as exensive quantity $\langle N \rangle = V \times \langle n \rangle$.

β::AbstractFloat

Inverse temperature.

V::Int64

System size.

U₀::AbstractFloat

Intensive energy scale.

α::AbstractFloat

Extensive energy scale $\alpha = N / U_0$.

c::AbstractFloat

Fraction of history to discard when calculating averages.

μₜ₊₁::AbstractFloat

The next chemical potential value for time t+1.

μ_bar::AbstractFloat

The forgetful mean of the chemical potential $\bar{\mu}_t = \frac{1}{L_t} \sum_{t' = \lceil t/c \rceil}^t \mu_{t'}$ where $L_t = t - \lceil t/c \rceil + 1$.

μ_var::AbstractFloat

The forgetful variance of the chemical potential.

N_bar::AbstractFloat

The forgetful mean of the total particle number $\bar{N}_t = \frac{1}{L_t} \sum_{t' = \lceil t/c \rceil}^t N_{t'}$.

N_var::AbstractFloat

The forgetful variance of the total particle number.

N²_bar::AbstractFloat

The forgetful mean of the square of total particle number $\overline{N^2}_t = \frac{1}{L_t} \sum_{t' = \lceil t/c \rceil}^t N^2_{t'}$.

κ_bar::AbstractFloat

The forgetful mean of the compressibility $\bar{\kappa}_t = \frac{1}{L_t} \sum_{t' = \lceil t/c \rceil}^t \kappa_{t'}$.

μ_traj::Vector{T} where T<:AbstractFloat

Sequence $\mu_t$ of chemical potential values.

N_traj::Vector{T} where T<:AbstractFloat

Sequence $N_t$ of total particle number values.

N²_traj::Vector{T} where T<:AbstractFloat

Sequence $N^2_t$ of total particle number squared values.

MuTuner.MuTunerLogger — Method

MuTunerLogger(n₀::T, β::T, V::Int, U₀::T=1.0, c::T=0.5, μ₀::T=0.0, record::Bool=false)
    where {T<:AbstractFloat}

Constructs an instance of MuTunerLogger.

Arguments

n₀::T: Target particle density $\langle n \rangle$.

β::T: Inverse temperature.

V::Int: System size.

U₀::T=1.0: Characteristic intensive energy scale.

c::T=0.5: Fraction of history to discard when calculating forgetful averages and variances.

μ₀::T=0.0: Initial guess for chemical potential.

MuTuner.replay — Method

replay(μtuner::MuTunerLogger{T}) where {T<:AbstractFloat}

Replay the chemical potential tuner to its current state, returning the time series for all relevant quantities. Note that this functions allocates all new arrays when it returns the time series for various quantities.

MuTuner.update! — Method

update!(μtuner::MuTunerLogger{T}, n::T, N²::T) where {T<:AbstractFloat}

Update the chemical potential given new measurements of the particle density n, and the total particle number squared N².

MuTuner.update_forgetful_mean — Method

update_forgetful_mean(x::AbstractVector{T}, x̄ₜ::T, c::T)
    where {T<:AbstractFloat}

Given the previous value of the forgetful mean x̄ₜ, calculate its updated value x̄ₜ₊₁ assuming that x[end] = xₜ₊₁ has already been appended to x. The oldest c fraction of values is discarded when calculating the forgetful mean.

MuTuner.update_forgetful_mv — Method

update_forgetful_mv(x::AbstractVector{T}, V̄ₜ::T, x̄ₜ::T, c::T)
    where {T<:AbstractFloat}

Given the previous value of the forgetful variance V̄ₜ and foregetul mean x̄ₜ, calculate their updated values V̄ₜ₊₁ and x̄ₜ₊₁, assuming that x[end] = xₜ₊₁ has already been appended to x. The oldest c fraction of values is discarded when calculating the forgetful mean and variance.

MuTuner.update_forgetful_var — Method

update_forgetful_var(x::AbstractVector{T}, V̄ₜ::T, x̄ₜ₊₁::T, x̄ₜ::T, c::T)
    where {T<:AbstractFloat}

Given the previous value for the forgetful variance V̄ₜ and foregetul mean x̄ₜ, and the updated value for the forgetful mean x̄ₜ₊₁, calculate the updated value V̄ₜ₊₁, assuming that x[end] = xₜ₊₁ has already been appended to x. The oldest c fraction of values is discarded when calculating the forgetful variance.